requirements.txt
flask
spacy
pdfminer.six
python-docx
scikit-learn
pandas
nltk